requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
django-crispy-forms==2.0
django-encrypted-model-fields==0.6.5
django-environ==0.10.0
docutils==0.19
Faker==15.3.3
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
python-dateutil==2.8.2
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
